+++ account_billing_hook_recompute_vline/README.md
@@ -1,0 +1,41 @@
+Billing Process
+===============
+
+In some countries, there is a customary practice for
+companies to collect money from their customers only once in a month.
+For example, the customer has 3 payments due in a given month, the vendor or
+billing company should group all the due AR Invoices in a document call
+Billing Document and issue it with all the invoices consolidated to the
+customer on the Billing Day. The customer will be paying based on the payable
+amount shown in Billing Document in the following month.
+
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/account-invoicing/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
+`here <https://github.com/OCA/account-invoicing/issues/new?body=module:%account_billing%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+
+Credits
+=======
+
+Contributors
+------------
+
+* Kitti U. <kittiu@ecosoft.co.th>
+
+Maintainer
+----------
+
+.. image:: http://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: http://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose mission is to support the collaborative development of Odoo features and promote its widespread use.
+
+To contribute to this module, please visit http://odoo-community.org.
