+++ account_billing_hook_recompute_vline/__init__.py
@@ -17,26 +17,8 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 #
-{
-    'name': "Billing Process",
-    'summary': "Group invoice as billing before payment",
-    'author': "Ecosoft,Odoo Community Association (OCA)",
-    'website': "http://ecosoft.co.th",
-    'category': 'Account',
-    'version': '0.1.0',
-    'depends': ['account',
-                'account_voucher',
-                'account_accountant',
-                'account_billing_hook_recompute_vline',
-                ],
-    'data': [
-        'data/account_billing_data.xml',
-        'data/account_billing_sequence.xml',
-        'data/account_billing_workflow.xml',
-        'security/ir.model.access.csv',
-        'views/account_billing.xml',
-        'views/voucher_payment_receipt_view.xml',
-        ],
-    'installable': True,
-}
+
+from . import account_voucher
+from . import account_voucher_hook_recompute_vline
+
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
